refactor(cli): move cobra-mutex helpers to shared/cli/, dedupe 3 call sites
hasExplicitLongFlag and cobraMutuallyExclusiveErrorMessage have no
functions-domain logic — they're generic cobra/argv-parsing primitives.
Move them to shared/cli/cobra-flag-groups.ts and point every independent
implementation of the same template at them instead of leaving duplicates
behind:

- dump.handler.ts had its own inline copy of the exact same message
  template (with its own Option-presence "changed" detection, kept as-is).
- gen/types/types.handler.ts had its own hasExplicitLongFlag AND its own
  message template (ensureMutuallyExclusive) — the latter had a real bug:
  it joined the violating-flag set in check order instead of sorting it,
  so `gen types --local --linked` printed "[local linked]" while the real
  Go binary (verified directly) prints "[linked local]". Fixed as a direct
  consequence of pointing it at the shared, already-correct formatter.

Also documents download's missing flag-conflict exit code in
SIDE_EFFECTS.md to match deploy's existing row.
