+++ apps/cli/src/shared/functions/download.ts
@@ -9,8 +9,10 @@ import type * as HttpClientResponse from "effect/unstable/http/HttpClientRespons
 import { Output } from "../output/output.service.ts";
 import {
   cobraMutuallyExclusiveErrorMessage,
-  FUNCTIONS_BUNDLER_MUTEX_GROUP,
   hasExplicitLongFlag,
+} from "../cli/cobra-flag-groups.ts";
+import {
+  FUNCTIONS_BUNDLER_MUTEX_GROUP,
   invalidFunctionSlugDetail,
   validateFunctionSlugMessage,
 } from "./functions.shared.ts";
